webservice/requirements.txt
pydantic
fastapi
mlflow
uvicorn
google-cloud-storage
python-dotenv